requirements.txt
flask
python-dotenv
pandas
google-generativeai
Flask-SQLAlchemy
Flask-Migrate
rq
redis